+++ CHANGELOG.md
@@ -1,6 +1,61 @@
-# Change Log
+# Changelog
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+
+### [3.0.2](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/compare/v3.0.1...v3.0.2) (2020-05-05)
+
+
+### Bug Fixes
+
+* return correct exit code when thresholds are not met (that works with the karma api) ([a0cbdd8](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/commit/a0cbdd80a2268586b62ef72e5541a36b1b46396a)), closes [#80](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/issues/80)
+
+### [3.0.1](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/compare/v3.0.0...v3.0.1) (2020-05-04)
+
+
+### Bug Fixes
+
+* set correct process exit code when thresholds are not met ([9ba3fe0](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/commit/9ba3fe0acf3ee778518c154a8dd050584f47bc13)), closes [#80](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/issues/80)
+
+## [3.0.0](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/compare/v2.1.1...v3.0.0) (2020-05-01)
+
+
+### ⚠ BREAKING CHANGES
+
+* node 6 and node 8 are no longer supported. To use this package you must upgrade to node 10 or higher.
+
+### Features
+
+* upgrade to latest istanbul api ([28cbbfb](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/commit/28cbbfb2cf8bd5b9533ceb489cc5047ff3def730))
+
+<a name="2.1.1"></a>
+## [2.1.1](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/compare/v2.1.0...v2.1.1) (2019-11-21)
+
+
+### Bug Fixes
+
+* add npm funding link ([ee9fb51](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/commit/ee9fb51))
+
+
+
+<a name="2.1.0"></a>
+# [2.1.0](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/compare/v2.0.6...v2.1.0) (2019-07-24)
+
+
+### Features
+
+* **config:** support istanbul-api instrumentation configuration ([c4f7a9c](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/commit/c4f7a9c)), closes [#73](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/issues/73)
+
+
+
+<a name="2.0.6"></a>
+## [2.0.6](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/compare/v2.0.5...v2.0.6) (2019-07-20)
+
+
+### Bug Fixes
+
+* get source code from sourceMapStore on write report ([619d90d](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/commit/619d90d)), closes [#72](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/issues/72)
+
+
 
 <a name="2.0.5"></a>
 ## [2.0.5](https://github.com/mattlewis92/karma-coverage-istanbul-reporter/compare/v2.0.4...v2.0.5) (2019-02-17)